# docs/openapi.yaml
openapi: 3.1.0
info:
  title: IMan Investment Platform API
  version: 1.0.0
  summary: Canonical production contract for the public website and administration portal.
  description: |
    JSON field names are camelCase. IDs are UUIDs and timestamps are RFC 3339 UTC instants.
    Public reads expose only published content. `/admin/**` requires a JWT plus the permission
    stated by each operation. A SUPER_ADMIN bypasses individual permission checks.
servers:
  - url: /api
    description: Same-origin deployment
...
security: []
paths:
...
  /admin/dashboard:
    get:
      tags: [Dashboard]
      operationId: getDashboard
      description: 'RBAC: dashboard.view'
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Dashboard aggregates
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Dashboard' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /admin/business-plans:
    get:
      tags: [Business plans]
      operationId: listBusinessPlans
      description: 'RBAC: business_plans.manage'
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Size' }
        - { $ref: '#/components/parameters/Search' }
        - { $ref: '#/components/parameters/AdminSort' }
        - { name: status, in: query, schema: { $ref: '#/components/schemas/BusinessPlanStatus' } }
        - { name: assignedToId, in: query, schema: { type: string, format: uuid } }
        - { name: submittedFrom, in: query, schema: { type: string, format: date-time } }
        - { name: submittedTo, in: query, schema: { type: string, format: date-time } }
      responses:
        '200':
          description: Matching plans
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BusinessPlanPage' }
  /admin/business-plans/{planId}:
    get:
      tags: [Business plans]
      operationId: getBusinessPlan
      description: 'RBAC: business_plans.manage'
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/PlanId' }]
      responses:
        '200':
          description: Plan detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BusinessPlan' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Business plans]
      operationId: archiveBusinessPlan
      description: 'RBAC: business_plans.manage. Soft-archives the record and retains audit/history.'
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/PlanId' }]
      responses:
        '204': { description: Archived }
  /admin/business-plans/{planId}/status:
    patch:
      tags: [Business plans]
      operationId: transitionBusinessPlan
      description: 'RBAC: business_plans.manage. Server validates the workflow transition.'
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/PlanId' }
        - { $ref: '#/components/parameters/IfMatch' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StatusTransitionRequest' }
      responses:
        '200':
          description: Updated plan
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BusinessPlan' }
        '409': { $ref: '#/components/responses/Conflict' }
  /admin/business-plans/{planId}/assignment:
    patch:
      tags: [Business plans]
      operationId: assignBusinessPlan
      description: 'RBAC: business_plans.manage'
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/PlanId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId]
              properties:
                userId: { type: [string, 'null'], format: uuid, description: Null unassigns. }
      responses:
        '200':
          description: Updated plan
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BusinessPlan' }
  /admin/business-plans/{planId}/notes:
    post:
      tags: [Business plans]
      operationId: addBusinessPlanNote
      description: 'RBAC: business_plans.manage'
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/PlanId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NoteCreate' }
      responses:
        '201':
          description: Created note
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Note' }
  /admin/business-plans/{planId}/history:
    get:
      tags: [Business plans]
      operationId: getBusinessPlanHistory
      description: 'RBAC: business_plans.manage'
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/PlanId' }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Size' }
      responses:
        '200':
          description: Immutable workflow history
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WorkflowEventPage' }
  /admin/business-plans/{planId}/documents/{documentId}:
    get:
      tags: [Business plans]
      operationId: downloadBusinessPlanDocument
      description: 'RBAC: business_plans.manage'
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/PlanId' }
        - { name: documentId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: File content
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/applicants:
    get:
      tags: [Applicants]
      operationId: listApplicants
      description: 'RBAC: applicants.manage'
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Size' }
        - { $ref: '#/components/parameters/Search' }
        - { $ref: '#/components/parameters/AdminSort' }
        - { name: status, in: query, schema: { $ref: '#/components/schemas/ApplicantStatus' } }
        - { name: jobId, in: query, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Matching applicants
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApplicantPage' }
  /admin/applicants/{applicantId}:
    get:
      tags: [Applicants]
      operationId: getApplicant
      description: 'RBAC: applicants.manage'
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/ApplicantId' }]
      responses:
        '200':
          description: Applicant detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Applicant' }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/applicants/{applicantId}/status:
    patch:
      tags: [Applicants]
      operationId: transitionApplicant
      description: 'RBAC: applicants.manage. Interview and offer transitions require their corresponding record.'
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/ApplicantId' }
        - { $ref: '#/components/parameters/IfMatch' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ApplicantTransitionRequest' }
      responses:
        '200':
          description: Updated applicant
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Applicant' }
        '409': { $ref: '#/components/responses/Conflict' }
  /admin/applicants/{applicantId}/notes:
    post:
      tags: [Applicants]
      operationId: addApplicantNote
      description: 'RBAC: applicants.manage'
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/ApplicantId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NoteCreate' }
      responses:
        '201':
          description: Created note
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Note' }
...
  /admin/portfolio:
    get:
      tags: [Portfolio administration]
      operationId: listAdminPortfolio
      description: 'RBAC: portfolio.manage'
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Size' }
        - { $ref: '#/components/parameters/Search' }
        - { $ref: '#/components/parameters/AdminSort' }
        - { name: status, in: query, schema: { $ref: '#/components/schemas/ContentStatus' } }
      responses:
        '200':
          description: Portfolio records
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PortfolioCompanyPage' }
    post:
      tags: [Portfolio administration]
      operationId: createPortfolioCompany
      description: 'RBAC: portfolio.manage'
      security: [{ bearerAuth: [] }]
      requestBody: { $ref: '#/components/requestBodies/PortfolioWrite' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PortfolioCompany' }
  /admin/portfolio/{id}:
    get:
      tags: [Portfolio administration]
      operationId: getPortfolioCompany
      description: 'RBAC: portfolio.manage'
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: Portfolio company
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PortfolioCompany' }
    put:
      tags: [Portfolio administration]
      operationId: updatePortfolioCompany
      description: 'RBAC: portfolio.manage'
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody: { $ref: '#/components/requestBodies/PortfolioWrite' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PortfolioCompany' }
    delete:
      tags: [Portfolio administration]
      operationId: deletePortfolioCompany
      description: 'RBAC: portfolio.manage'
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '204': { description: Deleted }
  /admin/jobs:
    get:
      tags: [Job administration]
      operationId: listAdminJobs
      description: 'RBAC: jobs.manage'
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Size' }
        - { $ref: '#/components/parameters/Search' }
        - { $ref: '#/components/parameters/AdminSort' }
        - { name: status, in: query, schema: { $ref: '#/components/schemas/JobStatus' } }
      responses:
        '200':
          description: Jobs
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JobPage' }
    post:
      tags: [Job administration]
      operationId: createJob
      description: 'RBAC: jobs.manage'
      security: [{ bearerAuth: [] }]
      requestBody: { $ref: '#/components/requestBodies/JobWrite' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Job' }
  /admin/jobs/{id}:
    get:
      tags: [Job administration]
      operationId: getAdminJob
      description: 'RBAC: jobs.manage'
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: Job
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Job' }
    put:
      tags: [Job administration]
      operationId: updateJob
      description: 'RBAC: jobs.manage'
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody: { $ref: '#/components/requestBodies/JobWrite' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Job' }
    delete:
      tags: [Job administration]
      operationId: deleteJob
      description: 'RBAC: jobs.manage; rejected while applicants reference the job unless it is archived.'
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '204': { description: Deleted }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /admin/media/{id}:
    delete:
      tags: [Media]
      operationId: deleteMedia
      description: 'RBAC: media.manage; returns 409 while referenced by published content.'
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '204': { description: Deleted }
        '409': { $ref: '#/components/responses/Conflict' }
  /admin/media/{id}/download:
    get:
      tags: [Media]
      operationId: downloadMedia
      description: 'RBAC: media.manage'
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200': { description: File, content: { application/octet-stream: { schema: { type: string, format: binary } } } }
...
  /admin/notifications/{id}:
    delete:
      tags: [Operations]
      operationId: deleteOwnNotification
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '204': { description: Deleted }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived access JWT issued by `/auth/login`.
  parameters:
...
    PlanId: { name: planId, in: path, required: true, schema: { type: string, format: uuid } }
...
    Page: { name: page, in: query, description: Zero-based page index., schema: { type: integer, minimum: 0, default: 0 } }
    Size: { name: size, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
    Search: { name: search, in: query, schema: { type: string, maxLength: 200 } }
    AdminSort: { name: sort, in: query, description: 'Single allow-listed field and direction.', schema: { type: string, default: 'createdAt,desc', pattern: '^[A-Za-z][A-Za-z0-9]*,(asc|desc)$' } }
...
  responses:
...
    Forbidden: { description: Authenticated but not authorized., content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
...
  schemas:
...
    BusinessPlanStatus: { type: string, enum: [DRAFT, SUBMITTED, REVIEWING, APPROVED, REJECTED, ARCHIVED] }
...